XML Schema with concrete connectives.
A minimal set of cardinalities must be lifted to get
better validation. An alternative XML Schema is included,
which is derived from the official by removing extensions
on the logical connectives and related constructs. This leads
to a new schema which is valid, and which validates all XML
rule files included in the distribution.